+++ angular/projects/lib/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bnqkl/web-component-angular
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bnqkl/web-component@1.7.0
+
 ## 2.0.6
 
 ### Patch Changes